Refactor authoring app, merge with inline-authoring-form
[#168140239]
